requirements.txt
apache-beam[gcp]>=2.50.0
google-cloud-storage>=2.0.0
google-cloud-bigquery>=3.0.0